[chore] #21 plan -> main.plan으로 폴더 이동
