+++ app/src/main/java/org/sopt/pingle/presentation/ui/main/plan/PlanDateTimeFragment.kt
@@ -1,12 +1,12 @@
-package org.sopt.pingle.presentation.ui.plan
+package org.sopt.pingle.presentation.ui.main.plan
 
 import android.os.Bundle
 import android.view.View
 import org.sopt.pingle.R
-import org.sopt.pingle.databinding.FragmentPlanOpenChattingBinding
+import org.sopt.pingle.databinding.FragmentPlanDateTimeBinding
 import org.sopt.pingle.util.base.BindingFragment
 
-class PlanOpenChattingFragment : BindingFragment<FragmentPlanOpenChattingBinding>(R.layout.fragment_plan_open_chatting) {
+class PlanDateTimeFragment : BindingFragment<FragmentPlanDateTimeBinding>(R.layout.fragment_plan_date_time) {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
     }
